Fiddling around when waiting for OCP restart

+++ countidon/countidon/CountidonItem.swift
@@ -19,12 +19,12 @@ class CountidonItem: NSObject, NSCoding {
   /**
       Time to countdown in seconds
    */
-  var timeToCountdown: NSInteger = 0
+  var timeToCountdown: Int = 0
   let maxTime: Double = 12.0
   var lastInterval = TimeInterval()
   var timer = Timer()
   var totalTime = TimeInterval()
-  var beepEvery: NSInteger = 0
+  var beepEvery: Int = 0
 
   
   var circleForegroundColor = UIColor(hue:1, saturation:1, brightness:0, alpha:1)
